What are those floaty things in your eye?
Have you ever noticed something swimming in your field of vision? It may look like a tiny worm or a transparent blob, and whenever you try to get a closer look, it disappears, only to reappear as soon as you shift your glance. But don't go rinsing out your eyes! What you are seeing is a common phenomenon known as a floater. The scientific name for these objects is Muscae volitantes, Latin for "flying flies," and true to their name, they can be somewhat annoying. But they're not actually bugs or any kind of external objects at all. Rather, they exist inside your eyeball. Floaters may seem to be alive, since they move and change shape, but they are not alive. Floaters are tiny objects that cast shadows on the retina, the light-sensitive tissue at the back of your eye. They might be bits of tissue, red blood cells, or clumps of protein. And because they're suspended within the vitreous humor, the gel-like liquid that fills the inside of your eye, floaters drift along with your eye movements, and seem to bounce a little when your eye stops. Floaters may be only barely distinguishable most of the time. They become more visible the closer they are to the retina, just as holding your hand closer to a table with an overhead light will result in a more sharply defined shadow. And floaters are particularly noticeable when you are looking at a uniform bright surface, like a blank computer screen, snow, or a clear sky, where the consistency of the background makes them easier to distinguish. The brighter the light is, the more your pupil contracts. This has an effect similar to replacing a large diffuse light fixture with a single overhead light bulb, which also makes the shadow appear clearer. There is another visual phenomenon that looks similar to floaters but is in fact unrelated. If you've seen tiny dots of light darting about when looking at a bright blue sky, you've experienced what is known as the blue field entoptic phenomenon. In some ways, this is the opposite of seeing floaters. Here, you are not seeing shadows but little moving windows letting light through to your retina. The windows are actually caused by white blood cells moving through the capillaries along your retina's surface. These leukocytes can be so large that they nearly fill a capillary causing a plasma space to open up in front of them. Because the space and the white blood cells are both more transparent to blue light than the red blood cells normally present in capillaries, we see a moving dot of light wherever this happens, following the paths of your capillaries and moving in time with your pulse. Under ideal viewing conditions, you might even see what looks like a dark tail following the dot. This is the red blood cells that have bunched up behind the leukocyte. Some science museums have an exhibit which consists of a screen of blue light, allowing you to see these blue sky sprites much more clearly than you normally would. While everybody's eyes experience these sort of effects, the number and type vary greatly. In the case of floaters, they often go unnoticed, as our brain learns to ignore them. However, abnormally numerous or large floaters that interfere with vision may be a sign of a more serious condition, requiring immediate medical treatment. But the majority of the time entoptic phenomena, such as floaters and blue sky sprites, are just a gentle reminder that what we think we see depends just as much on our biology and minds as it does on the external world.